Update for bitmap, image on scaling, transparancy,   wxCHECKs and no more dependance of wxBitmap on wxImage   in its API
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@1117 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/listctrl.cpp
@@ -2004,7 +2004,7 @@ void wxListMainWindow::DeleteItem( long index )
 
 void wxListMainWindow::DeleteColumn( int col )
 {
-  wxCHECK_RET( col < m_columns.GetCount(),
+  wxCHECK_RET( col < (int)m_columns.GetCount(),
                "attempting to delete inexistent column in wxListView" );
 
   m_dirty = TRUE;
